requirements.txt
numpy==1.24.3
pandas==2.0.3
pyarrow==12.0.1
scipy==1.10.1
matplotlib==3.8.4
psutil==5.9.8
tqdm==4.66.5
pyyaml==6.0.2
orjson==3.10.7
